+++ src/tools/toThousands.js
@@ -1,6 +1,6 @@
 function toThousands(v = 0, figures) {
   if (typeof +v === 'number' && !isNaN(v) ) {
-    const arr = `${v}`.split('.'); // 将数字进行整数位和小数位的分隔
+    const arr = `${+v}`.split('.'); // 将数字进行整数位和小数位的分隔
     const integer = arr[0].replace(/\B(?=(\d{3})+(?!\d))/g, ',');
     const decimal = (figures || arr[1]) && (arr[1] || '').padEnd(figures, 0);