+++ docs/changeLog.md
@@ -7,6 +7,12 @@ order: 1
 
 遵循 [Semantic Versioning 2.0.0](https://semver.org/lang/zh-CN/) 语义化版本规范。
 
+### 🚀 2.11.3
+
+- TdUpload
+  - fileTypes 支持 video、audio 文件类型
+- 🐞 tools.toThousands.currency 修复当传入空字符串时，显示错误的问题
+
 ### 🚀 2.11.2
 
 - TdUpload